requirements.txt
pandas
streamlit
google-api-python-client
google-auth
google-auth-oauthlib
openai
serpapi
python-dotenv
